+++ semgrep-core/tests/js/deep_stmt3.js
@@ -1,0 +1,4 @@
+//ERROR: match
+function f() {
+    console.log();
+}