+++ semgrep/semgrep/config_resolver.py
@@ -1,25 +1,26 @@
 import os
-import sys
-import tarfile
-import tempfile
 import time
 from pathlib import Path
-from typing import Any
 from typing import Dict
 from typing import List
 from typing import Optional
+
+from ruamel.yaml import YAMLError
 
 from semgrep.constants import DEFAULT_CONFIG_FILE
 from semgrep.constants import DEFAULT_CONFIG_FOLDER
 from semgrep.constants import DEFAULT_SEMGREP_CONFIG_NAME
 from semgrep.constants import ID_KEY
+from semgrep.constants import PLEASE_FILE_ISSUE_TEXT
 from semgrep.constants import RULES_KEY
 from semgrep.constants import SEMGREP_USER_AGENT
 from semgrep.constants import YML_EXTENSIONS
+from semgrep.error import SemgrepError
+from semgrep.rule_lang import parse_yaml_preserve_spans
+from semgrep.rule_lang import YamlTree
 from semgrep.util import debug_print
 from semgrep.util import is_url
 from semgrep.util import print_error
-from semgrep.util import print_error_exit
 from semgrep.util import print_msg
 
 IN_DOCKER = "SEMGREP_IN_DOCKER" in os.environ
@@ -31,25 +32,32 @@ TEMPLATE_YAML_URL = (
 )
 
 RULES_REGISTRY = {
-    "r2c": "https://semgrep.live/c/r/all",
+    "r2c": "https://semgrep.live/c/p/r2c",
 }
 DEFAULT_REGISTRY_KEY = "r2c"
 
 
-def manual_config(pattern: str, lang: str) -> Dict[str, Any]:
+def manual_config(pattern: str, lang: str) -> Dict[str, Optional[YamlTree]]:
     # TODO remove when using sgrep -e ... -l ... instead of this hacked config
+    pattern_tree = parse_yaml_preserve_spans(pattern, filename=None)
+    error_span = parse_yaml_preserve_spans(
+        f"Semgrep bug generating manual config {PLEASE_FILE_ISSUE_TEXT}", filename=None
+    ).span
     return {
-        "manual": {
-            RULES_KEY: [
-                {
-                    ID_KEY: "-",
-                    "pattern": pattern,
-                    "message": pattern,
-                    "languages": [lang],
-                    "severity": "ERROR",
-                }
-            ]
-        }
+        "manual": YamlTree.wrap(
+            {
+                RULES_KEY: [
+                    {
+                        ID_KEY: "-",
+                        "pattern": pattern_tree,
+                        "message": pattern,
+                        "languages": [lang],
+                        "severity": "ERROR",
+                    }
+                ]
+            },
+            span=error_span,
+        )
     }
 
 
@@ -65,7 +73,7 @@ def adjust_for_docker(in_precommit: bool = False) -> None:
     # change into this folder so that all paths are relative to it
     if IN_DOCKER and not IN_GH_ACTION and not in_precommit:
         if not Path(REPO_HOME_DOCKER).exists():
-            print_error_exit(
+            raise SemgrepError(
                 f'you are running semgrep in docker, but you forgot to mount the current directory in Docker: missing: -v "${{PWD}}:{REPO_HOME_DOCKER}"'
             )
     if Path(REPO_HOME_DOCKER).exists():
@@ -82,36 +90,32 @@ def indent(msg: str) -> str:
 
 def parse_config_at_path(
     loc: Path, base_path: Optional[Path] = None
-) -> Dict[str, Optional[Dict[str, Any]]]:
+) -> Dict[str, Optional[YamlTree]]:
     config_id = str(loc)
     if base_path:
         config_id = str(loc).replace(str(base_path), "")
     try:
         with loc.open() as f:
-            return parse_config_string(config_id, f.read())
+            return parse_config_string(config_id, f.read(), str(loc))
     except FileNotFoundError:
         print_error(f"YAML file at {loc} not found")
         return {str(loc): None}
 
 
 def parse_config_string(
-    config_id: str, contents: str
-) -> Dict[str, Optional[Dict[str, Any]]]:
-    import yaml  # here for faster startup times
-
+    config_id: str, contents: str, filename: Optional[str]
+) -> Dict[str, Optional[YamlTree]]:
     try:
-        return {config_id: yaml.safe_load(contents)}
-    except yaml.parser.ParserError as se:
-        print_error(f"Invalid yaml file {config_id}:\n{indent(str(se))}")
-        return {config_id: None}
-    except yaml.scanner.ScannerError as se:
+        data = parse_yaml_preserve_spans(contents, filename)
+        return {config_id: data}
+    except YAMLError as se:
         print_error(f"Invalid yaml file {config_id}:\n{indent(str(se))}")
         return {config_id: None}
 
 
 def parse_config_folder(
     loc: Path, relative: bool = False
-) -> Dict[str, Optional[Dict[str, Any]]]:
+) -> Dict[str, Optional[YamlTree]]:
     configs = {}
     for l in loc.rglob("*"):
         # Allow manually specified paths with ".", but don't auto-expand them
@@ -136,7 +140,7 @@ def _is_hidden_config(loc: Path) -> bool:
 
 def load_config_from_local_path(
     location: Optional[str] = None,
-) -> Dict[str, Optional[Dict[str, Any]]]:
+) -> Dict[str, Optional[YamlTree]]:
     """
         Return config file(s) as dictionary object
     """
@@ -158,23 +162,26 @@ def load_config_from_local_path(
             elif loc.is_dir():
                 return parse_config_folder(loc)
             else:
-                print_error_exit(f"config location `{loc}` is not a file or folder!")
+                raise SemgrepError(f"config location `{loc}` is not a file or folder!")
         else:
             addendum = ""
             if IN_DOCKER:
                 addendum = " (since you are running in docker, you cannot specify arbitary paths on the host; they must be mounted into the container)"
-            print_error_exit(
+            raise SemgrepError(
                 f"unable to find a config; path `{loc}` does not exist{addendum}"
             )
     raise Exception
 
 
-def download_config(config_url: str) -> Dict[str, Optional[Dict[str, Any]]]:
+def download_config(config_url: str) -> Dict[str, Optional[YamlTree]]:
     import requests  # here for faster startup times
 
-    DOWNLOADING_MESSAGE = "downloading config..."
-    SCANNING_MESSAGE = "scanning code...     "  # needs to have same number of chars as DOWNLOADING_MESSAGE
+    DOWNLOADING_MESSAGE = f"downloading config..."
+    SCANNING_MESSAGE = "scanning code...\033[K"
     debug_print(f"trying to download from {config_url}")
+    print_msg(
+        f"using config from {config_url}. See https://semgrep.live/registry for more options."
+    )
     print_msg(DOWNLOADING_MESSAGE, end="\r")
     headers = {"User-Agent": SEMGREP_USER_AGENT}
 
@@ -182,24 +189,24 @@ def download_config(config_url: str) -> Dict[str, Optional[Dict[str, Any]]]:
         r = requests.get(config_url, stream=True, headers=headers, timeout=10)
         if r.status_code == requests.codes.ok:
             content_type = r.headers.get("Content-Type")
-            if content_type and "text/plain" in content_type:
+            yaml_types = [
+                "text/plain",
+                "application/x-yaml",
+                "text/x-yaml",
+                "text/yaml",
+                "text/vnd.yaml",
+            ]
+            if content_type and any((ct in content_type for ct in yaml_types)):
                 print_msg(SCANNING_MESSAGE)
-                return parse_config_string("remote-url", r.content.decode("utf-8"))
-            elif content_type and content_type == "application/x-gzip":
-                with tempfile.TemporaryDirectory() as fname:
-                    with tarfile.open(fileobj=r.raw, mode="r:gz") as tar:
-                        tar.extractall(fname)
-                    extracted = Path(fname)
-                    for path in extracted.iterdir():
-                        # get first folder in extracted folder (this is how GH does it)
-                        print_msg(SCANNING_MESSAGE)
-                        return parse_config_folder(path, relative=True)
+                return parse_config_string(
+                    "remote-url", r.content.decode("utf-8"), filename=None
+                )
             else:
-                print_error_exit(
+                raise SemgrepError(
                     f"unknown content-type: {content_type} returned by config url: {config_url}. Can not parse"
                 )
         else:
-            print_error_exit(
+            raise SemgrepError(
                 f"bad status code: {r.status_code} returned by config url: {config_url}"
             )
     except Exception as e:
@@ -207,7 +214,7 @@ def download_config(config_url: str) -> Dict[str, Optional[Dict[str, Any]]]:
     return {config_url: None}
 
 
-def resolve_config(config_str: Optional[str]) -> Dict[str, Optional[Dict[str, Any]]]:
+def resolve_config(config_str: Optional[str]) -> Dict[str, Optional[YamlTree]]:
     """ resolves if config arg is a registry entry, a url, or a file, folder, or loads from defaults if None"""
     start_t = time.time()
     if config_str is None:
@@ -228,7 +235,7 @@ def generate_config() -> None:
 
     # defensive coding
     if Path(DEFAULT_CONFIG_FILE).exists():
-        print_error_exit(
+        raise SemgrepError(
             f"{DEFAULT_CONFIG_FILE} already exists. Please remove and try again"
         )
     try:
@@ -250,6 +257,5 @@ def generate_config() -> None:
         with open(DEFAULT_CONFIG_FILE, "w") as template:
             template.write(template_str)
             print_msg(f"Template config successfully written to {DEFAULT_CONFIG_FILE}")
-            sys.exit(0)
     except Exception as e:
-        print_error_exit(str(e))
+        raise SemgrepError(str(e))