+++ semgrep/semgrep/semgrep_main.py
@@ -1,3 +1,4 @@
+import json
 import sys
 from pathlib import Path
 from typing import Any
@@ -9,94 +10,96 @@ from typing import Tuple
 import semgrep.config_resolver
 from semgrep.autofix import apply_fixes
 from semgrep.constants import DEFAULT_CONFIG_FILE
-from semgrep.constants import ID_KEY
 from semgrep.constants import OutputFormat
 from semgrep.constants import RULES_KEY
 from semgrep.core_runner import CoreRunner
+from semgrep.error import FINDINGS_EXIT_CODE
+from semgrep.error import INVALID_CODE_EXIT_CODE
+from semgrep.error import InvalidPatternNameError
+from semgrep.error import InvalidRuleSchemaError
+from semgrep.error import MISSING_CONFIG_EXIT_CODE
+from semgrep.error import SemgrepError
 from semgrep.output import build_output
 from semgrep.rule import Rule
+from semgrep.rule_lang import YamlTree
 from semgrep.rule_match import RuleMatch
-from semgrep.semgrep_types import InvalidRuleSchema
 from semgrep.semgrep_types import YAML_ALL_VALID_RULE_KEYS
 from semgrep.semgrep_types import YAML_MUST_HAVE_KEYS
 from semgrep.util import debug_print
-from semgrep.util import FINDINGS_EXIT_CODE
-from semgrep.util import INVALID_CODE_EXIT_CODE
 from semgrep.util import is_url
-from semgrep.util import MISSING_CONFIG_EXIT_CODE
 from semgrep.util import print_error
-from semgrep.util import print_error_exit
 from semgrep.util import print_msg
 
 MISSING_RULE_ID = "no-rule-id"
 
 
-def validate_single_rule(config_id: str, rule: Dict[str, Any]) -> bool:
+def validate_single_rule(config_id: str, rule_yaml: YamlTree) -> Optional[Rule]:
     """
         Validate that a rule dictionary contains all necessary keys
         and can be correctly parsed
     """
-    rule_id_err_msg = f'{rule.get("id", MISSING_RULE_ID)}'
+    rule: Dict[str, Any] = rule_yaml.value  # type: ignore
+    rule_id_err_msg = f'{rule.get("id", MISSING_RULE_ID).unroll()}'
     rule_keys = set(rule.keys())
     if not rule_keys.issuperset(YAML_MUST_HAVE_KEYS):
         missing_keys = YAML_MUST_HAVE_KEYS - rule_keys
         print_error(
             f"{config_id} is missing required keys {missing_keys} at rule id {rule_id_err_msg}"
         )
-        return False
+        return None
     if not rule_keys.issubset(YAML_ALL_VALID_RULE_KEYS):
         extra_keys = rule_keys - YAML_ALL_VALID_RULE_KEYS
         print_error(
-            f"{config_id} has invalid rule key {extra_keys} at rule id {rule_id_err_msg}, can only have: {YAML_ALL_VALID_RULE_KEYS}"
+            f"{config_id} has an invalid top-level rule key {extra_keys} at rule id {rule_id_err_msg}, can only have: {sorted(YAML_ALL_VALID_RULE_KEYS)}"
         )
-        return False
+        return None
     try:
-        _ = Rule.from_json(rule).expression
-    except InvalidRuleSchema as ex:
+        return Rule.from_yamltree(rule_yaml)
+    except (InvalidPatternNameError, InvalidRuleSchemaError) as ex:
         print_error(
             f"{config_id}: inside rule id {rule_id_err_msg}, pattern fields can't look like this: {ex}"
         )
-        return False
-    try:
-        _ = Rule.from_json(rule).globs
-    except InvalidRuleSchema as ex:
-        print_error(
-            f"{config_id}: inside rule id {rule_id_err_msg}, path fields can't look like this: {ex}"
-        )
-        return False
-
-    return True
+        return None
 
 
 def validate_configs(
-    configs: Dict[str, Optional[Dict[str, Any]]]
-) -> Tuple[Dict[str, Any], Dict[str, Any]]:
+    configs: Dict[str, Optional[YamlTree]]
+) -> Tuple[Dict[str, List[Rule]], Dict[str, Any]]:
     """
         Take configs and separate into valid and invalid ones
     """
-    errors = {}
-    valid = {}
-    for config_id, config in configs.items():
-        if not config:
-            errors[config_id] = config
+    errors: Dict[str, Any] = {}
+    valid: Dict[str, Any] = {}
+    for config_id, config_yaml_tree in configs.items():
+        if not config_yaml_tree:
+            errors[config_id] = config_yaml_tree
             continue
-        if RULES_KEY not in config:
+        config = config_yaml_tree.value
+        if not isinstance(config, dict):
+            print_error(f"{config_id} was not a mapping")
+            errors[config_id] = config_yaml_tree.unroll()
+            continue
+        rules = config.get(RULES_KEY)  # type: ignore
+        if rules is None:
             print_error(f"{config_id} is missing `{RULES_KEY}` as top-level key")
-            errors[config_id] = config
+            errors[config_id] = config_yaml_tree.unroll()
             continue
-        rules = config.get(RULES_KEY)
         valid_rules = []
         invalid_rules = []
-        for rule in rules:  # type: ignore
-            if validate_single_rule(config_id, rule):
+        for rule_dict in rules.value:
+            rule = validate_single_rule(config_id, rule_dict)
+            if rule:
                 valid_rules.append(rule)
             else:
-                invalid_rules.append(rule)
+                invalid_rules.append(rule_dict)
 
         if invalid_rules:
-            errors[config_id] = {**config, "rules": invalid_rules}
+            errors[config_id] = {
+                **config_yaml_tree.unroll_dict(),
+                "rules": invalid_rules,
+            }
         if valid_rules:
-            valid[config_id] = {**config, "rules": valid_rules}
+            valid[config_id] = valid_rules
     return valid, errors
 
 
@@ -118,18 +121,15 @@ def convert_config_id_to_prefix(config_id: str) -> str:
     return prefix
 
 
-def rename_rule_ids(valid_configs: Dict[str, Any]) -> Dict[str, Any]:
+def rename_rule_ids(valid_configs: Dict[str, List[Rule]]) -> Dict[str, List[Rule]]:
     transformed = {}
-    for config_id, config in valid_configs.items():
-        rules = config.get(RULES_KEY, [])
-        transformed_rules = [
-            {
-                **rule,
-                ID_KEY: f"{convert_config_id_to_prefix(config_id)}{rule.get(ID_KEY, MISSING_RULE_ID)}",
-            }
+    for config_id, rules in valid_configs.items():
+        transformed[config_id] = [
+            rule.with_id(
+                f"{convert_config_id_to_prefix(config_id)}{rule.id or MISSING_RULE_ID}"
+            )
             for rule in rules
         ]
-        transformed[config_id] = {**config, RULES_KEY: transformed_rules}
     return transformed
 
 
@@ -144,7 +144,7 @@ def post_output(output_url: str, output: str) -> None:
         r = requests.post(output_url, data=output, timeout=10)
         debug_print(f"posted to {output_url} and got status_code:{r.status_code}")
     except requests.exceptions.Timeout:
-        print_error_exit(f"posting output to {output_url} timed out")
+        raise SemgrepError(f"posting output to {output_url} timed out")
 
 
 def save_output(destination: str, output: str) -> None:
@@ -162,16 +162,19 @@ def save_output(destination: str, output: str) -> None:
             fout.write(output)
 
 
-def get_config(generate_config: bool, pattern: str, lang: str, config: str) -> Any:
+def get_config(
+    generate_config: bool, pattern: str, lang: str, config: str
+) -> Tuple[Dict[str, List[Rule]], Dict[str, Any]]:
     # first check if user asked to generate a config
     if generate_config:
         semgrep.config_resolver.generate_config()
+        sys.exit(0)
 
     # let's check for a pattern
     elif pattern:
         # and a language
         if not lang:
-            print_error_exit("language must be specified when a pattern is passed")
+            raise SemgrepError("language must be specified when a pattern is passed")
 
         # TODO for now we generate a manual config. Might want to just call semgrep -e ... -l ...
         configs = semgrep.config_resolver.manual_config(pattern, lang)
@@ -181,7 +184,7 @@ def get_config(generate_config: bool, pattern: str, lang: str, config: str) -> A
 
     # if we can't find a config, use default r2c rules
     if not configs:
-        print_error_exit(
+        raise SemgrepError(
             f"No config given and {DEFAULT_CONFIG_FILE} was not found. Try running with --help to debug or if you want to download a default config, try running with --config r2c"
         )
 
@@ -192,19 +195,8 @@ def get_config(generate_config: bool, pattern: str, lang: str, config: str) -> A
     return valid_configs, invalid_configs
 
 
-def flatten_configs(transformed_configs: Dict[str, Any]) -> List[Rule]:
-    return [
-        Rule.from_json(rule)
-        for config in transformed_configs.values()
-        for rule in config.get(RULES_KEY, [])
-    ]
-
-
-def should_exclude_this_path(path: Path) -> bool:
-    """
-        Return true if "test" or "example" are anywhere in path
-    """
-    return any("test" in p or "example" in p for p in path.parts)
+def flatten_configs(transformed_configs: Dict[str, List[Rule]]) -> List[Rule]:
+    return [rule for rules in transformed_configs.values() for rule in rules]
 
 
 def main(
@@ -219,8 +211,8 @@ def main(
     include_dir: List[str],
     exclude: List[str],
     exclude_dir: List[str],
-    exclude_tests: bool,
     json_format: bool,
+    debugging_json: bool,
     sarif: bool,
     output_destination: str,
     quiet: bool,
@@ -236,13 +228,15 @@ def main(
     output_format = OutputFormat.TEXT
     if json_format:
         output_format = OutputFormat.JSON
+    elif debugging_json:
+        output_format = OutputFormat.JSON_DEBUG
     elif sarif:
         output_format = OutputFormat.SARIF
 
     if invalid_configs and strict:
-        print_error_exit(
+        raise SemgrepError(
             f"run with --strict and there were {len(invalid_configs)} errors loading configs",
-            MISSING_CONFIG_EXIT_CODE,
+            code=MISSING_CONFIG_EXIT_CODE,
         )
 
     if not no_rewrite_rule_ids:
@@ -267,13 +261,13 @@ def main(
         )
 
         if len(valid_configs) == 0:
-            print_error_exit(
+            raise SemgrepError(
                 f"no valid configuration file found ({len(invalid_configs)} configs were invalid)",
-                MISSING_CONFIG_EXIT_CODE,
+                code=MISSING_CONFIG_EXIT_CODE,
             )
 
     # actually invoke semgrep
-    rule_matches_by_rule, semgrep_errors = CoreRunner(
+    rule_matches_by_rule, debug_steps_by_rule, semgrep_errors = CoreRunner(
         allow_exec=dangerously_allow_arbitrary_code_execution_from_rules,
         jobs=jobs,
         exclude=exclude,
@@ -282,36 +276,15 @@ def main(
         include_dir=include_dir,
     ).invoke_semgrep(targets, all_rules)
 
-    if exclude_tests:
-        ignored_in_tests = 0
-        filtered_findings_by_rule = {}
-        for rule, rule_matches in rule_matches_by_rule.items():
-            filtered = []
-            for rule_match in rule_matches:
-                if should_exclude_this_path(Path(rule_match.path)):
-                    ignored_in_tests += 1
-                else:
-                    filtered.append(rule_match)
-            filtered_findings_by_rule[rule] = filtered
-        rule_matches_by_rule = filtered_findings_by_rule
-        if ignored_in_tests > 0:
-            print_error(
-                f"warning: ignored {ignored_in_tests} results in tests due to --exclude-tests option"
-            )
-
-    for finding in semgrep_errors:
-        print_error(f"semgrep: {finding['path']}: {finding['check_id']}")
-
-    if strict and len(semgrep_errors):
-        print_error_exit(
-            f"run with --strict and {len(semgrep_errors)} errors occurred during semgrep run; exiting",
-            INVALID_CODE_EXIT_CODE,
-        )
+    if output_format == OutputFormat.TEXT:
+        for error in semgrep_errors:
+            print_error(pretty_error(error))
 
     output = handle_output(
         rule_matches_by_rule,
         semgrep_errors,
         output_format,
+        debug_steps_by_rule,
         quiet,
         output_destination,
         exit_on_error,
@@ -319,13 +292,35 @@ def main(
     if autofix:
         apply_fixes(rule_matches_by_rule)
 
+    if strict and len(semgrep_errors):
+        raise SemgrepError(
+            f"run with --strict and {len(semgrep_errors)} errors occurred during semgrep run; exiting",
+            code=INVALID_CODE_EXIT_CODE,
+        )
+
     return output
+
+
+def pretty_error(error: Dict[str, Any]) -> str:
+    try:
+        if {"path", "start", "end", "extra"}.difference(error.keys()) == set():
+            header = f"{error['extra']['message']}\n--> {error['path']}:{error['start']['line']}"
+            line_1 = error["extra"]["line"]
+            start_col = error["start"]["col"]
+            line_2 = " " * (start_col - 1) + "^"
+            line_3 = f"= note: If the code is correct, this could be a semgrep bug -- please help us fix this by filing an an issue at https://semgrep.dev"
+            return "\n".join([header, line_1, line_2, line_3])
+        else:
+            return f"semgrep-core error: {json.dumps(error,indent=2)}"
+    except KeyError:
+        return f"semgrep-core error: {json.dumps(error, indent=2)}"
 
 
 def handle_output(
     rule_matches_by_rule: Dict[Rule, List[RuleMatch]],
     semgrep_errors: List[Any],
     output_format: OutputFormat,
+    debug_steps_by_rule: Dict[Rule, List[Dict[str, Any]]],
     quiet: bool = False,
     output_destination: Optional[str] = None,
     exit_on_error: bool = False,
@@ -338,7 +333,12 @@ def handle_output(
     ]
 
     output = build_output(
-        rule_matches, rules, semgrep_errors, output_format, not output_destination
+        rule_matches,
+        debug_steps_by_rule,
+        rules,
+        semgrep_errors,
+        output_format,
+        not output_destination,
     )
     if not quiet:
         if output: