+++ semgrep/tests/e2e/test_semgrep_core_parse_error.py
@@ -1,0 +1,17 @@
+from subprocess import CalledProcessError
+
+import pytest
+
+
+@pytest.mark.parametrize(
+    "filename", ["invalid_python.py",],
+)
+def test_rule_parser__failure__error_messages(run_semgrep_in_tmp, snapshot, filename):
+    with pytest.raises(CalledProcessError) as excinfo:
+        run_semgrep_in_tmp(
+            config="rules/eqeq-python.yaml",
+            target_name=f"bad/{filename}",
+            output_format="text",
+            stderr=True,
+        )
+    snapshot.assert_match(excinfo.value.output, "error.txt")